Annotation Checkboxes - Story 26 complete
Added checkboxes for the boat name and boat speed annotations. Also added interaction with them and the Annotations Button, which now acts to always try to show all of the annotations unless they are both already showing, in which case it will turn them both off.  Clicking the checkboxes will also correctly change the annotations button text.

(Recommit, as this was reverted at some point in master)

#story[26]

Author: Jonathan Travaille (jtr44)