Story 30/46/47 Merging - Combining parts from all the stories to work in unison
Utilised dependency injection to enable the passing of information between stream client, parsers and the race itself.
Merged the branches for stories 30 and 46 (with story 47 also done on the 46 branch).
Developed processes to ensure that the data has been read in by the parser to set how the corresponding objects act.

#story[954, 950, 951] #pair[msi52, osr13, asv15]