Story 34:  Completed Story 34. Sparklines load in properly and have a css file to allow for changes.
Fixed data addition to the sparklines.
Added a CSS file that sets the background colour as transparent, and in code the axis are hidden.
Reduced the timer to check and update boat position twice every second.
Fixed the adjustments of the sidepanel with the listview such that both the listview and sparklines are still shown.
Removed the horizontal scrollbar haunting the tableview.

#story[866]  #pair[msi52, tjg73]
Author: Matt Simpson msi52