Story 30: Fixed the race clock to display time to or from the race start time.
Altered expected race start time variable from LocalDateTime to long.
Retrieved UTC offset to get the time from the race xml file to convert the expected race start time to a long.
If the application starts mid race, it displays the time since the race started.
If prior to the race starting, a countdown is started that counts down to the race starting.

#story[866]  #pair[msi52, asv15, jtr44]