Story 30: Fixing boat order in race.
Battled with fixing the positioning based on a comparator and sorting how it all fit together.
In the end, based it on the leg index of the boat. Boats that are nearer the end of the pack ( lower leg index), are at the start of the list, and the boats to the front of the pack are at the end of the list ()the right).
When a boat is sorted, if the leg is equal to or less than, it is on the left, otherwise, to the right.

#story[954] #pair[tjg44, msi52]