Story 30: More fixing boat order in race. Fixed comparator. Fixed sparkline display. Added more CSS.
Attempted to alter comparator for a more accurate positioning system.
Changed the sizing of the spakline chart so that it is always the same size after resizing.
Changed colouration of boats, now boats are given a set colour that matches the colour in the sparkline chart.
Removed sparkline chart legend.

#story[954] #pair[tjg73, msi52]