Story 59: Altered race class in mock files to dynamically generate wind speed and direction instead of reading from xml file.
The race class now determines its own data for the wind direction. It starts by instantiating a random number in the range of 0, 360. Then when the getter for the wind direction is called, the wind direction may change based upon a probability.

The race class now determines its own data for the wind speed. The first gust of wind is selected by choosing a random value between 5 and 40 knots,  and from there the wind value can only be changed to either 5 knots above or 5 knots below that original value

#story[#1027]