Added visual signal of collisions to boats. Added manual testing for collisions. Fixed the message data being sent upon collision occurring.
Changed way that the collision data is retained client side. The client side race now only cares about collisions occurring with the boat under the control of that client.
Control circle around a boat fills in a red colour to indicate a collision. The length of time that a control circle is filled is controlled using the frames of the animation timer in the race controller.

#story[1256]